Make it so on Ruby 1.9
We have to do a few things here. First, encoding pragmas. Second, we move to minitest from test-spec which went belly-up when it has acquired bacon-its-evil-twin. Third, we introduce Bundler into the mix and let the updated Jeweler use it. The whole shebang should still work on 1.8.7 which we cannot guarantee yet due to a botched sqlite3 dependency, but this is something we are going to fix very soon. Tests green on 1.9.3-p0 with Rails 3.1