The gemspec needs to require date from stdlib for Date.today to work on all Rubies
Apparently Date.today is NOT available without requiring it first. Exhibit:

    bigbuk ~/Code/apps/tracksperanto-web(master) $ /Users/julik/.rvm/rubies/ruby-1.9.3-p429/bin/irb
    1.9.3p429 :001 > Date.methods.grep /today/
     => []

This means that on servers the Bundler require step will fail for the checkout of 'peiji_san', and apps will not boot. This was not a problem on 1.9.3-p0 but is now (on p125 and p429). Instead of changing the system Ruby install it might be a good idea to just require the date from stdlib and be done with it.